JpushDone

+++ app/src/main/java/com/zhongchuangtiyu/denarau/Utils/DensityUtil.java
@@ -11,7 +11,7 @@ public class DensityUtil
     /**
      * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
      */
-    public static int dp2px(Context context, float dpValue) {
+    public static int dp2px(Context context, int dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * scale + 0.5f);
     }
